refactor: move category DTOs to dedicated `dto` subpackage
This commit reorganizes the project structure by moving category-related Data Transfer Objects (DTOs) into a specific `dto` subpackage, ensuring consistency with other modules like `auth` and `transaction`.

- **DTO Migration**: Relocated `CategoryDto.kt`, `CategoryDtoWithTransactions.kt`, and `GetCategoriesResponse.kt` to the `id.my.sendiko.fintrack.category.data.dto` package.
- **Import Updates**:
    - Updated `ApiService.kt` and `KtorClient.kt` to reference the new location of `GetCategoriesResponse`.
    - Updated `TransactionsDtoWithWallet.kt` to reference the new location of `CategoryDto`.
