+++ shared/src/commonMain/kotlin/id/my/sendiko/fintrack/category/data/dto/CategoryDtoWithTransactions.kt
@@ -1,10 +1,11 @@
-package id.my.sendiko.fintrack.category.data
+package id.my.sendiko.fintrack.category.data.dto
 
-import kotlinx.serialization.SerialName
+import id.my.sendiko.fintrack.transaction.data.dto.TransactionsDto
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.SerialName
 
 @Serializable
-data class CategoryDto(
+data class CategoryDtoWithTransactions(
 
 	@SerialName("createdAt")
 	val createdAt: String,
@@ -14,6 +15,9 @@ data class CategoryDto(
 
 	@SerialName("id")
 	val id: String,
+
+	@SerialName("transactions")
+	val transactions: List<TransactionsDto>,
 
 	@SerialName("userId")
 	val userId: String,
