+++ shared/src/commonMain/kotlin/id/my/sendiko/fintrack/category/data/dto/GetCategoriesResponse.kt
@@ -1,4 +1,4 @@
-package id.my.sendiko.fintrack.category.data
+package id.my.sendiko.fintrack.category.data.dto
 
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.SerialName
