+++ shared/src/commonMain/kotlin/id/my/sendiko/fintrack/core/network/KtorClient.kt
@@ -7,7 +7,7 @@ import id.my.sendiko.fintrack.auth.login.data.dto.LoginRequest
 import id.my.sendiko.fintrack.auth.login.data.dto.LoginResponse
 import id.my.sendiko.fintrack.auth.register.data.dto.RegisterRequest
 import id.my.sendiko.fintrack.auth.register.data.dto.RegisterResponse
-import id.my.sendiko.fintrack.category.data.GetCategoriesResponse
+import id.my.sendiko.fintrack.category.data.dto.GetCategoriesResponse
 import id.my.sendiko.fintrack.core.network.utils.DataError
 import id.my.sendiko.fintrack.core.network.utils.Result
 import id.my.sendiko.fintrack.transaction.data.dto.GetTransactionsResponse
